Move esp.web.util.latex to esp.utils.latex
